+++ pages/index.vue
@@ -3,6 +3,7 @@
     <CardComponent class="test" :character="character" />
     <button class="button" @click="newRandomPnj()">nouveau pnj</button>
     <div class="distribution"><array-distribution :data="data" /></div>
+    <stats />
   </div>
 </template>
 
@@ -14,7 +15,7 @@ const data = ref(null);
 
 import CardComponent from '/components/card-component.vue';
 onMounted(async () => {
-  // newRandomPnj()
+  newRandomPnj()
   data.value = await getEthnicDistributionByNationality();
 });
 
@@ -41,11 +42,37 @@ async function newRandomPnj() {
 }
 
 .button {
-  width: 150px;
+  width: auto;
   margin: 50px 0;
+  padding: 10px 20px;
+  background-color: $main-color; // Couleur de fond rouge
+  color: white; // Couleur du texte blanc
+  border: none;
+  border-radius: $spacing-xxs; // Coins arrondis
+  cursor: pointer;
+  transition: all 0.3s ease; // Effet de transition
+  font-family: $font-family-dosis;
+  text-transform: uppercase;
+  letter-spacing: $spacing-xs;
+
+  &:hover {
+    background-color: $secondary-color; // Couleur de fond quand survolé
+  }
+
+  &:active {
+    transform: scale(0.98); // Réduction de la taille quand cliqué
+  }
+
+  &:disabled {
+    background-color: $dark-gray-color; // Couleur de fond quand désactivé
+    cursor: not-allowed;
+  }
 }
+
 
 .distribution {
   width: 50%;
 }
+
+
 </style>