+++ services/apiService.ts
@@ -14,7 +14,17 @@ export const getRandomCharacter = async () => {
 
 export const getEthnicDistributionByNationality = async () => {
   try {
-    const response = await axios.get('http://localhost:3001/global/ethnic-distribution-by-nationality');
+    const response = await axios.get('http://localhost:3001/global/information-table/ethnic-distribution-by-nationality');
+    return response.data;
+  } catch (error) {
+    console.error('Erreur lors de l\'appel à l\'API :', error);
+    return null;
+  }
+};
+
+export const getStats = async () => {
+  try {
+    const response = await axios.get('http://localhost:3001/global/information-table/stats');
     return response.data;
   } catch (error) {
     console.error('Erreur lors de l\'appel à l\'API :', error);